Backend/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
pyodbc==5.1.0
Werkzeug==2.2.2
google-cloud-storage==2.5.0
